Linux 2.2.20pre3
o	FPU/ptrace corruption fixes			(Victor Zandy)
o	Resync belkin usb serial with 2.4		(Greg Kroah-Hartmann)
o	Resync digiport usb serial with 2.4		(Greg Kroah-Hartmann)
o	Rsync empeg usb serial with 2.4			(Greg Kroah-Hartmann)
o	Resync ftdi_sio against 2.4			(Greg Kroah-Hartmann)
o	Bring keyscan usb back into line with 2.4	(Greg Kroah-Hartmann)
o	Resync keyspan_pda usb with 2.4			(Greg Kroah-Hartmann)
o	Resync omninet usb with 2.4.5			(Greg Kroah-Hartmann)
o	Resync usb-serial driver with 2.4.5		(Greg Kroah-Hartmann)
o	Resync visor usb driver with 2.4.5		(Greg Kroah-Hartmann)
o	Rsync whiteheat driver with 2.4.5		(Greg Kroah-Hartmann)
o	Add edgeport USB serial				(Greg Kroah-Hartmann)
o	Add mct_u232 USB serial				(Greg Kroah-Hartmann)
o	Update usb storage device list		(Stas Bekman, Kaz Sasayma)
o	Bring usb acm driver into line with 2.4.5	(Greg Kroah-Hartmann)
o	Bring bluetooth driver into line with 2.4.5	(Greg Kroah-Hartmann)
o	Bring dabusb driver into line with 2.4.5	(Greg Kroah-Hartmann)
o	Bring usb dc2xx driver into line with 2.4.5	(Greg Kroah-Hartmann)
o	Bring mdc800 usb driver into line with 2.4.5	(Greg Kroah-Hartmann)
o	Bring rio driver into line with 2.4.5		(Greg Kroah-Hartmann)
o	Bring USB scanner drivers into line with 2.4.5	(Greg Kroah-Hartmann)
o	Update ov511 driver to match 2.4.5		(Greg Kroah-Hartmann)
o	Update PCIIOC ioctls (esp for sparc)		(Dave Miller)
o	General sparc bugfixes				(Dave Miller)
o	Fix possible oops in fbmem ioctls		(Dave Miller)
o	Fix reboot/halt bug on "Alcor" Alpha boxes	(Tom Vier)
o	Update osst driver 				(Willem Riede)
o	Fix syncppp negotiation bug			(Bob Dunlop)
o	SMBfs bug fixes from 2.4 series			(Urban Widmark)
o	3ware IDE raid driver updates			(Adam Radford)
o	Fix incorrect use of bitops on non long types	(Dave Miller)
o	Fix reboot/halt bug on 'Miata' Alpha boxes	(Tom Vier)
o	Update Tim Waugh's contact info			(Tim Waugh)
o	Add TIOCGSERIAL to sun serial on PCI sparc32	(Lars Kellogg-Stedman)
o	ov511 check user data more carefully		(Marc McClelland)
o	Fix netif_wake_queue compatibility macro	(Andi Kleen)

+++ drivers/block/ide-floppy.c
@@ -125,7 +125,7 @@ typedef struct idefloppy_packet_command_s {
 	byte *current_position;			/* Pointer into the above buffer */
 	void (*callback) (ide_drive_t *);	/* Called when this packet command is completed */
 	byte pc_buffer[IDEFLOPPY_PC_BUFFER_SIZE];	/* Temporary buffer */
-	unsigned int flags;			/* Status/Action bit flags */
+	unsigned long flags;			/* Status/Action bit flags */
 } idefloppy_pc_t;
 
 /*
@@ -259,7 +259,7 @@ typedef struct {
 	idefloppy_flexible_disk_page_t flexible_disk_page;	/* Copy of the flexible disk page */
 	int wp;							/* Write protect */
 
-	unsigned int flags;			/* Status/Action flags */
+	unsigned long flags;			/* Status/Action flags */
 } idefloppy_floppy_t;
 
 /*
